+++ src/main/java/apple/discord/clover/api/ApiModule.java
@@ -1,8 +1,8 @@
 package apple.discord.clover.api;
 
+import apple.discord.clover.api.character.CharacterController;
 import apple.discord.clover.api.meta.ping.PingController;
-import apple.discord.clover.api.player.activity.PlayerController;
-import apple.discord.clover.api.player.character.CharacterController;
+import apple.discord.clover.api.player.overview.PlayerController;
 import apple.lib.modules.AppleModule;
 import apple.lib.modules.configs.factory.AppleConfigLike;
 import io.javalin.Javalin;
