+++ src/main/java/apple/discord/clover/api/character/CharacterController.java
@@ -1,7 +1,8 @@
-package apple.discord.clover.api.player.character;
+package apple.discord.clover.api.character;
 
 import apple.discord.clover.api.base.ApiController;
-import apple.discord.clover.api.player.activity.PlayerRequest;
+import apple.discord.clover.api.character.request.CharacterRequest;
+import apple.discord.clover.api.character.response.CharacterStatsListResponse;
 import apple.discord.clover.database.query.character.CharacterTermsQuery;
 import io.javalin.Javalin;
 import io.javalin.http.Context;
@@ -20,7 +21,7 @@ public class CharacterController extends ApiController {
 
     private void characterTerms(Context ctx) {
         String queryParam = new String(Base64.getUrlDecoder().decode(ctx.queryParam("search")));
-        PlayerRequest request = gson().fromJson(queryParam, PlayerRequest.class);
+        CharacterRequest request = gson().fromJson(queryParam, CharacterRequest.class);
         CharacterStatsListResponse response = CharacterTermsQuery.queryCharacterTerms(request);
         ctx.json(stringify(response));
     }
