+++ src/main/java/apple/discord/clover/api/character/request/CharacterRequest.java
@@ -1,0 +1,10 @@
+package apple.discord.clover.api.character.request;
+
+import apple.discord.clover.api.base.TermRequest;
+import java.util.UUID;
+
+public class CharacterRequest extends TermRequest {
+
+    public UUID player;
+
+}
