+++ src/main/java/apple/discord/clover/api/player/overview/request/PlayerRequest.java
@@ -1,0 +1,9 @@
+package apple.discord.clover.api.player.overview.request;
+
+import apple.discord.clover.api.base.TermRequest;
+import java.util.UUID;
+
+public class PlayerRequest extends TermRequest {
+
+    public UUID player;
+}
