+++ src/main/java/apple/discord/clover/api/player/overview/response/PlayerResponse.java
@@ -1,4 +1,4 @@
-package apple.discord.clover.api.player.activity;
+package apple.discord.clover.api.player.overview.response;
 
 import apple.discord.clover.database.activity.DPlaySession;
 import java.time.Instant;
